requirements.txt
pandas
qrcode
pillow
requests
firebase_admin